Add --stop and --keep flags

+++ run.sh
@@ -1,5 +1,4 @@
 #!/bin/sh
-
 set -e
 
 if [ "$OS" = "Windows_NT" ]; then
@@ -13,6 +12,23 @@ else
 	*) echo "Unsupported platform"; exit 1 ;;
 	esac
 fi
+
+# Flags parsing
+STOP=0
+KEEP=0
+for arg in "$@"; do
+	case $arg in
+	-K|--keep)
+		# Keep the unzipped executable after running
+		KEEP=1
+		;;
+	-S|--stop)
+		# Don't run the executable after unzipping (also keeps it)
+		STOP=1
+		KEEP=1
+		;;
+	esac
+done
 
 # Fetch the latest release download URL
 REPO="Xevion/spotify-quickauth"
@@ -32,5 +48,9 @@ curl -Lso $EXECUTABLE.tar.gz $DOWNLOAD_URL
 tar -xvf $EXECUTABLE.tar.gz $EXECUTABLE 1>/dev/null
 rm $EXECUTABLE.tar.gz
 chmod +x $EXECUTABLE
-trap "rm -f $EXECUTABLE" INT EXIT
-./$EXECUTABLE $@
+if [ "$KEEP" -eq 0 ]; then
+	trap "rm -f $EXECUTABLE" INT EXIT
+fi
+if [ "$STOP" -eq 0 ]; then
+	./$EXECUTABLE $@
+fi
